Refactor: EditorBuiltinResources script interface now exported via script binding generator  - Also added paths to 3D scene icons  - Updated 'bsf' version

+++ Source/EditorManaged/Inspectors/LightProbeVolumeInspector.cs
@@ -89,7 +89,7 @@ namespace BansheeEditor
             removeProbeButton.OnClick += () =>
             {
                 if (LightProbeVolumeNodeHandles.SelectedNode != uint.MaxValue)
-                    lpv.RemoveProbe(LightProbeVolumeNodeHandles.SelectedNode);
+                    lpv.RemoveProbe((int)LightProbeVolumeNodeHandles.SelectedNode);
 
                 MarkAsModified();
                 ConfirmModify();
